requirements.txt
paho-mqtt>=1.5.0
rpyc>=4.1.5
requests>=2.22.0
appdirs>=1.4.3
